Allow omitting allowed when passing allowedHelp
Closes #25

Use the keys from `allowedHelp` as the default for `allowed` to avoid
requiring callers to specify the same thing twice.

This is potentially breaking if a caller was already only passing
`allowedHelp` which impacts the usage string but does not restrict the
values in practice. This is unlikely to have been intentional since the
usage string would not be correct, but it there is a possibility that
some library was doing the disallowed check outside of the parsing and
was intending to use the argument only to influence the usage text.